cambios2.0

+++ api/src/main/java/pide/devuelve/repository/PrestamoRepository.java
@@ -2,5 +2,5 @@ package pide.devuelve.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import pide.devuelve.entidades.Prestamo;
 
-public interface PrestamoRepository extends JpaRepository<Prestamo, Long> {
+public interface PrestamoRepository extends JpaRepository<Prestamo, Integer> {
 }
