+++ api/src/main/java/pide/devuelve/repository/RecursoRepository.java
@@ -2,5 +2,5 @@ package pide.devuelve.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import pide.devuelve.entidades.Recurso;
 
-public interface RecursoRepository extends JpaRepository<Recurso, Long> {
+public interface RecursoRepository extends JpaRepository<Recurso, Integer> {
 }
